Added gamelist json and update functionality
Made a json file that stores game objects from an api call. It is setup to check the date in the json file to the current day's date, and if outdate, it will update the json file. Once updated it will call from the local json file, reducing load times greatly.